requirement.txt
sqlalchemy
passlib[bcrypt]
aiofiles
asyncpg
PyJWT
python-dotenv